+++ CHANGELOG.md
@@ -11,6 +11,7 @@ CentOS-7 7.5.1804 x86_64 - MySQL 5.7 Community Server.
 - Updates source image to [2.5.0](https://github.com/jdeathe/centos-ssh/releases/tag/2.5.0).
 - Updates and restructures Dockerfile.
 - Adds placeholder replacement of `RELEASE_VERSION` docker argument to systemd service unit template.
+- Adds consideration for event lag into test cases for unhealthy health_status events.
 - Removes use of `/etc/services-config` paths.
 - Removes code from configuration file `/etc/mysqld-bootstrap.conf`.
 - Removes X-Fleet section from etcd register template unit-file.